pinger-bot/requirements.txt
python-telegram-bot==13.7
requests==2.28.2
schedule==1.1.0